+++ src/lustre/ui_internal/dom.gleam
@@ -1,7 +1,7 @@
 // IMPORTS ---------------------------------------------------------------------
 
 import gleam/dynamic.{type Dynamic}
-import gleam/dynamic/decode.{type Decoder}
+import gleam/dynamic/decode.{type DecodeError, type Decoder}
 
 // TYPES -----------------------------------------------------------------------
 
@@ -15,12 +15,6 @@ pub fn assigned_elements(_slot: HtmlElement) -> List(HtmlElement)
 @external(javascript, "./dom.ffi.mjs", "tag")
 pub fn tag(element: HtmlElement) -> String
 
-@external(javascript, "./dom.ffi.mjs", "text_content")
-pub fn text_content(element: HtmlElement) -> String
-
-@external(javascript, "./dom.ffi.mjs", "children")
-pub fn children(element: HtmlElement) -> List(HtmlElement)
-
 ///
 pub fn attribute(element: HtmlElement, name: String) -> Result(String, Nil) {
   do_attribute(element, name)
@@ -28,6 +22,16 @@ pub fn attribute(element: HtmlElement, name: String) -> Result(String, Nil) {
 
 @external(javascript, "./dom.ffi.mjs", "get_attribute")
 fn do_attribute(_element: HtmlElement, _name: String) -> Result(String, Nil)
+
+///
+///
+pub fn property(
+  element: HtmlElement,
+  name: String,
+  decoder: Decoder(a),
+) -> Result(a, List(DecodeError)) {
+  decode.run(as_dynamic(element), decode.at([name], decoder))
+}
 
 // DECODERS --------------------------------------------------------------------
 
@@ -45,6 +49,9 @@ fn is_element(_value: Dynamic) -> Bool
 
 @external(javascript, "../../../gleam_stdlib/gleam/function.mjs", "identity")
 fn as_element(_: Dynamic) -> HtmlElement
+
+@external(javascript, "../../../gleam_stdlib/gleam/function.mjs", "identity")
+fn as_dynamic(_: HtmlElement) -> Dynamic
 
 @external(javascript, "./dom.ffi.mjs", "make_fallback_element")
 fn make_fallback_element() -> HtmlElement
