+++ src/lustre/ui_internal/value.gleam
@@ -1,0 +1,8 @@
+// TYPES -----------------------------------------------------------------------
+
+///
+///
+pub type Value(a) {
+  Controlled(value: a)
+  Uncontrolled(value: a, default: a, touched: Bool)
+}
